OPHAKTKEH-209 Add styling for clerk autocomplete
