+++ src/main/reactjs/src/styles/abstracts/_loops.scss
@@ -1,0 +1,5 @@
+@each $value in $flex-grow-values {
+  .flex-grow-#{$value} {
+    flex-grow: $value;
+  }
+}
