+++ src/main/reactjs/src/styles/abstracts/_variables.scss
@@ -27,3 +27,5 @@ $breakpoint-md: 80rem;
 $breakpoint-lg: 96rem;
 $breakpoint-xl: 128rem;
 $breakpoint-xxl: 160rem;
+
+$flex-grow-values: 1, 2, 3, 4;
