+++ src/main/reactjs/src/styles/styles.scss
@@ -5,6 +5,7 @@
 @import 'abstracts/mixins';
 @import 'abstracts/variables';
 @import 'abstracts/maps';
+@import 'abstracts/loops';
 @import 'abstracts/common/spacing';
 @import 'abstracts/common/layout';
 @import 'abstracts/common/typograph';
